Split lagrangian_engine into engine1-4 modules
Refactor the monolithic lagrangian_engine into four focused modules: engine1_seeds (seeding kernels), engine2_pde (differential operators + PDE evolution/relaxation), engine3_observables (energy/trackers/sampling) and engine4_render (rendering helpers). Added those new files and removed lagrangian_engine.py, updating _launcher.py and several experiment scripts to import and call the new modules (e.g. seed_*, evolve_psi, relax_director_step, compute_energyH/F, sample/update rendering routines). This is a separation-of-concerns SoC refactor to clarify responsibilities and prepare for further PDE/observable/render enhancements.
