+++ openwave/xperiments/m5_liquid_crystal/research/sandbox_v2/m5_2_kg_defect_survival.py
@@ -39,7 +39,9 @@ if str(REPO_ROOT) not in sys.path:
 
 from openwave.common import constants  # noqa: E402
 from openwave.xperiments.m5_liquid_crystal import medium  # noqa: E402
-from openwave.xperiments.m5_liquid_crystal import lagrangian_engine as lagrange  # noqa: E402
+from openwave.xperiments.m5_liquid_crystal import engine1_seeds as seeds  # noqa: E402
+from openwave.xperiments.m5_liquid_crystal import engine2_pde as pde  # noqa: E402
+from openwave.xperiments.m5_liquid_crystal import engine3_observables as observables  # noqa: E402
 
 # ================================================================
 # CONFIG
@@ -62,11 +64,11 @@ def seed_single_hedgehog(wf):
     centers[0, 2] = wf.nz // 2
     signs[0] = +1
     D_quarter = float(DOMAIN_QUARTER_FRACTION * max(wf.nx, wf.ny, wf.nz))
-    lagrange.seed_hedgehog(wf, centers, signs, D_quarter, 1)
+    seeds.seed_hedgehog(wf, centers, signs, D_quarter, 1)
     for _ in range(N_RELAX):
         cfl_bound = (wf.dx_am**2) / 6.0
         tau = 0.4 * cfl_bound
-        lagrange.relax_director_step(wf, tau, centers, signs, 1)
+        pde.relax_director_step(wf, tau, centers, signs, 1)
         wf.psi_am.copy_from(wf.psi_new_am)
         wf.psi_prev_am.copy_from(wf.psi_new_am)
     return wf.nx // 2, wf.ny // 2, wf.nz // 2
@@ -77,16 +79,16 @@ def propagate_and_measure(wf, c_amrs, dt_rs, m_freq_rs, center):
     samples = []  # list of (step, Q, |psi|_avg)
     # Initial measurement
     psi_np = wf.psi_am.to_numpy()
-    Q0 = lagrange.compute_winding_number(psi_np, center, WINDING_RADIUS)
+    Q0 = observables.compute_winding_number(psi_np, center, WINDING_RADIUS)
     psi_norm = np.linalg.norm(psi_np, axis=-1).mean()
     samples.append((0, Q0, psi_norm))
 
     for step in range(1, N_PROPAGATE_STEPS + 1):
-        lagrange.evolve_psi(wf, c_amrs, dt_rs, m_freq_rs, 0.0)
+        pde.evolve_psi(wf, c_amrs, dt_rs, m_freq_rs, 0.0)
         wf.swap_buffers()
         if step % SAMPLE_EVERY == 0 or step == N_PROPAGATE_STEPS:
             psi_np = wf.psi_am.to_numpy()
-            Q = lagrange.compute_winding_number(psi_np, center, WINDING_RADIUS)
+            Q = observables.compute_winding_number(psi_np, center, WINDING_RADIUS)
             psi_norm = np.linalg.norm(psi_np, axis=-1).mean()
             samples.append((step, Q, psi_norm))
